exmove: set default center to (0,0)
exmove is inconsist between scale/rotate and normal move.
For scale and rotate, center is recomputed, and cx/cy is applied (50% if not set)
For normal move, center is left empty, which means the default center is (0,0)

To fix this inconsistency, set default center to always be zero.

Applied this change to the workflow example.
evey (ex)move now centered at (0,0)