Show face icon in history only at talk
Previously check page/line break to decide whether to print face icon.
But this approach does not work when high speed clicking or skipping.

With this fix, param talk should only be used when a different character start to talk.
Otherwise, use another action param to add animation to characters